Update 'Owned by' label for Workflow Progress component
