+++ src/ggrc/assets/assets.yaml
@@ -82,7 +82,6 @@ dashboard-js-template-files:
   - dashboard/internav_list.mustache
   - dashboard/info/info.mustache
   - dashboard/info/get_started.mustache
-  - dashboard/info/wf_owner.mustache
   - access_groups/modal_content.mustache
   - data_assets/modal_content.mustache
   - documents/active_column.mustache
