+++ src/ggrc/assets/javascripts/components/inline/inline-aggregate-field.mustache
@@ -1,0 +1,10 @@
+{{!
+  Copyright (C) 2017 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+}}
+
+<div rel="tooltip" data-original-title="{{resultStr}}" class="inline-aggregate-field">
+  <span>
+    {{aggregateValue}}
+  </span>
+</div>
