--- src/ggrc/assets/mustache/dashboard/info/wf_owner.mustache
@@ -1,25 +1,0 @@
-{{!
-    Copyright (C) 2017 Google Inc.
-    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-}}
-
-{{#instance}}   
-<!--p class="oneline"-->
-  {{^if_equals instance.kind "Backlog"}}
-    Owned by: 
-    {{#with_mapping 'authorizations' instance}}
-      {{#each authorizations}}
-      <span>
-        {{#using role=instance.role}}
-          {{#if_equals role.name 'WorkflowOwner'}}
-            {{#using contact=instance.person}}
-              {{{renderLive '/static/mustache/people/popover.mustache' person=contact}}}
-            {{/using}}
-          {{/if_equals}}
-        {{/using}}
-      </span>
-      {{/each}}
-    {{/with_mapping}}
-  {{/if_equals}}
-<!--/p-->
-{{/instance}}
